#4 LocalDateTIme 사용을 위한 JavaTimeModule 의존성 추가, register Test 코드 작성 완료

+++ ex05/src/main/java/ecount/quotation/mapper/QuotationMapper.java
@@ -11,11 +11,11 @@ public interface QuotationMapper {
 
 	List<QuotationDTO> findAllQuot(); // 다중 조회
 
-	QuotationDTO findQuotByAccNum(Long acc_number); // 단일 조회
+	QuotationDTO findQuotByAccNum(int acc_number); // 단일 조회
 
 	int insertQuot(QuotationDTO quot);
 
 	int updateQuot(QuotationDTO quot);
 
-	int deleteQuot(QuotationDTO quot); // 데이터를 삭제하지 않고 deleted_at 부여
+	int deleteQuot(QuotationDTO quot); // 데이터를 삭제하지 않고 delete_at 칼럼에 값 입력만
 }
